+++ archive/sqlite3/run-sqlite.sh
@@ -26,16 +26,16 @@
 #SBATCH --export=ALL
 #SBATCH --output="/mnt/titan/users/j.boom/logs/R-%x-%j.log"
 #SBATCH --error="/mnt/titan/users/j.boom/errors/R-%x-%j.error"
-#SBATCH --time=200:15:00
 #SBATCH --partition=all
 
 convert_dump(){
     # The convert_dump function:
     #     This function converts an sql dump to a sqlite database.
     #     https://www.rebasedata.com/convert-mysql-to-sqlite-online
+    source /home/j.boom/miniconda3/bin/activate base
     /mnt/titan/users/j.boom/tool-testing/mysql2sqlite/mysql2sqlite \
-        /mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/fathmm.v2.3.SQL \
-        | sqlite3 /mnt/titan/users/j.boom/tool-testing/vep/vep_grch37/plugins_data/fathmm.v2.3
+        /mnt/titan/users/j.boom/r-analysis/vep/plugins_data/fathmm.v2.3.SQL \
+        | sqlite3 /mnt/titan/users/j.boom/r-analysis/vep/plugins_data/fathmm.v2.3
 }
 
 main() {